+++ src/index.ts
@@ -1,5 +1,4 @@
 import AbstractDB from './abstract-db'
-import type { DBConfig, DBConnection } from './abstract-db'
 import AbstractQuery from './abstract-query'
 import FilteredQuery from './filtered-query'
 import InsertQuery from './insert-query'
@@ -12,12 +11,12 @@ import Relation from './relation'
 import QueryClauseGroup from './query-clause-group'
 import QueryClause from './query-clause'
 import ORMHelpers from './helpers'
-import { OrangeDatabaseError, OrangeDatabaseInputError, OrangeDatabaseQueryError, OrangeDatabaseUnexpectedResultError } from './errors'
-import type { JoinType } from './types'
+import { OrangeDatabaseError, OrangeDatabaseInputError, OrangeDatabaseQueryError, OrangeDatabaseUnexpectedResultError, OrangeDatabaseUsageError } from './errors'
 
 const transaction = async <T>(functionality: () => Promise<T>): Promise<T> => {
   return await AbstractQuery.db.transaction(functionality)
 }
 
-export { AbstractDB, AbstractQuery, FilteredQuery, InsertQuery, SelectQuery, UpdateQuery, DeleteQuery, ActiveRecord, RawSQL, Relation, QueryClauseGroup, QueryClause, ORMHelpers, transaction, OrangeDatabaseError, OrangeDatabaseInputError, OrangeDatabaseQueryError, OrangeDatabaseUnexpectedResultError }
-export type { JoinType, DBConfig, DBConnection }
+export { AbstractDB, AbstractQuery, FilteredQuery, InsertQuery, SelectQuery, UpdateQuery, DeleteQuery, ActiveRecord, RawSQL, Relation, QueryClauseGroup, QueryClause, ORMHelpers, transaction, OrangeDatabaseError, OrangeDatabaseInputError, OrangeDatabaseQueryError, OrangeDatabaseUnexpectedResultError, OrangeDatabaseUsageError }
+export type { DBConfig, DBConnection } from './abstract-db'
+export type { ClauseOperand, DeleteOptions, FieldFunctionObject, FullTextClauseFn, IActiveRecordConstructor, IActiveRecordInstance, IQueryBuilder, JoinType, OrderFieldObject, OrderSpec, RelationMode, SelectFields, SelectOptions, UpdateOptions, WhereOperator } from './types'